requirements.txt
# Pinned major ranges for lab/paper reproducibility (use `pip freeze` on a known-good machine for exact pins).
pupil-labs-realtime-api>=1.6,<2
streamlit>=1.28,<2
moviepy>=2.0,<3
imageio-ffmpeg>=0.4.0
filelock>=3.12.0
